github webhook create/delete complete, and hooked up to snap activate/deactivate

+++ src/snap/snap-dal.js
@@ -14,6 +14,7 @@
 const database = require('../data/database');
 const dbconstants = require('../data/database-constants');
 const engine = require('./snap-engine');
+const providers = require('../providers/providers');
 
 /* 
  * A snap definition is specified as follows:
@@ -31,18 +32,47 @@ const engine = require('./snap-engine');
 // activate a snap into the user's environment
 exports.activateSnap = async (userId, snapId, params = null) => {
   try {
+    // parse snapId (which is in account/name format)
+    const [user, name] = snapId.split('/');
+    if (!user || !name) {
+      return { message: `activateSnap: snapId ${snapId} not in user/name format` };
+    }
+
+    // get the snap structure
+    const snap = await database.getDocument(user, dbconstants.snapsCollection, name); 
+    if (!snap || !snap.trigger) {
+      return { message: `activateSnap: could not find snap ${user}/${name}` };
+    }
+
+    // get the provider for the trigger
+    const provider = providers.getProvider(snap.trigger);
+
     // active snap ID is current timestamp
     const activeSnapId = new Date().getTime().toString();
-    const record = {
+    const activeSnap = {
       activeSnapId: activeSnapId,
       userId: userId,
       snapId: snapId,
+      provider: provider.provider,
       params: params
     }
-    await database.storeDocument(userId, dbconstants.activeSnapsCollection, activeSnapId, record);
+
+    // create the snap trigger
+    const triggerData = await provider.createTrigger(userId, activeSnapId, params);
+    if (!triggerData) {
+      return { message: 'could not activate snap' };
+    }
+
+    // add the trigger data to the activesnap record
+    activeSnap.triggerData = triggerData;
+
+    // store the activated snap information
+    await database.storeDocument(userId, dbconstants.activeSnapsCollection, activeSnapId, activeSnap);
+
+    return { message: 'success' };
   } catch (error) {
     console.log(`activateSnap: caught exception: ${error}`);
-    return null;
+    return { message: 'could not activate snap' };
   }
 }
 
@@ -77,10 +107,22 @@ exports.createSnap = async (userId, definition, private = false) => {
 // deactivate a snap in the user's environment
 exports.deactivateSnap = async (userId, activeSnapId) => {
   try {
+    // get the active snap structure
+    const activeSnap = await database.getDocument(userId, dbconstants.activeSnapsCollection, activeSnapId);
+    if (!activeSnap) {
+      return { message: `could not find active snap ID ${activeSnapId}`};
+    }
+
+    // delete the snap trigger
+    const provider = providers.getProvider(activeSnap.provider);
+    await provider.deleteTrigger(userId, activeSnap.triggerData);
+
+    // delete the active snap from the database
     await database.removeDocument(userId, dbconstants.activeSnapsCollection, activeSnapId);
+    return { message: 'success' };
   } catch (error) {
     console.log(`deactivateSnap: caught exception: ${error}`);
-    return null;
+    return { message: `deactivateSnap error: ${error.message}`};
   }
 }
